# docs/openapi/combined.yaml
openapi: 3.0.3
info:
  title: Voice AI Platform API
  version: 1.0.0
  description: >
    Unified public API (via api-gateway :8080). All endpoints require
    `Authorization: Bearer <api_key>`. Live per-service interactive docs at each
    service's /docs (FastAPI auto-generated).
servers:
  - url: http://localhost:8080
security: [ { bearerAuth: [] } ]
paths:
  /v1/tts:
    post:
      summary: Synthesize speech
      requestBody:
        content:
          application/json:
            schema: { type: object, required: [text], properties: { text: {type: string}, voice: {type: string} } }
      responses:
        "200": { description: WAV audio; X-First-Chunk-Ms header = latency, content: { audio/wav: {} } }
...
  /v1/calls:
    post:
      summary: Originate outbound call
      requestBody: { content: { application/json: { schema: { type: object, required: [to], properties: { to: {type: string}, agent_prompt: {type: string} } } } } }
      responses: { "200": { description: "{call_id, status}" } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
...